Initial integration of ExoPlayer Initial integration of Hilt DI Library Base structure of packages created Basic UI implemented Code Cleanup Added dependencies (ExoPlayer, Hilt, Activity-ktx)
