+++ app/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
     package="com.overplay.blackmagicshinedemo">
 
     <application
+        android:name=".App"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -10,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.BlackMagicShineDemo">
         <activity
-            android:name=".FullscreenActivity"
+            android:name=".presentation.MainActivity"
             android:configChanges="orientation|keyboardHidden|screenSize"
             android:exported="true"
             android:label="@string/app_name"
