+++ app/src/main/java/com/overplay/blackmagicshinedemo/App.kt
@@ -1,0 +1,7 @@
+package com.overplay.blackmagicshinedemo
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class App : Application()
