+++ app/src/main/java/com/overplay/blackmagicshinedemo/presentation/MainViewModel.kt
@@ -1,0 +1,11 @@
+package com.overplay.blackmagicshinedemo.presentation
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class MainViewModel @Inject constructor() : ViewModel() {
+
+    //TODO implement media player instance and functions for transport control actions
+}
