+++ build.gradle
@@ -3,6 +3,7 @@ plugins {
     id 'com.android.application' version '7.1.3' apply false
     id 'com.android.library' version '7.1.3' apply false
     id 'org.jetbrains.kotlin.android' version '1.6.20' apply false
+    id 'dagger.hilt.android.plugin' version '2.41' apply false
 }
 
 task clean(type: Delete) {
